+++ tests/facet/RewardsFacet.test.ts
@@ -1,6 +1,6 @@
 import { assert, createMockedFunction, clearStore, test, describe, afterEach, dataSourceMock, beforeEach } from "matchstick-as/assembly/index";
 import { createApp, createBadge, createERC20Token, getTestBadgeTokenEntity, badgeMintedLegacy, mintERC20TokenAction, mintERC20TokenMission, transferBadge, transferERC20TokenMission, badgeMinted, erc721Minted, updatedBaseURI } from "../utils";
-import { TEST_ACTIONMETADATA_ENTITY_TYPE, TEST_ACTION_ENTITY_TYPE, TEST_APPFUNGIBLETOKEN_ENTITY_TYPE, TEST_APP_ID, TEST_BADGETOKEN_ENTITY_TYPE, TEST_BADGETOKEN_ID, TEST_MISSIONFUNGIBLETOKEN_ENTITY_TYPE, TEST_MISSIONMETADATA_ENTITY_TYPE, TEST_MISSION_ENTITY_TYPE, TEST_TOKEN_DECIMALS, TEST_TOKEN_ID, TEST_TOKEN_MINTED_ACTION_ID, TEST_TOKEN_MINTED_MISSION_ID, TEST_TOKEN_MINTED_URI, TEST_TOKEN_NAME, TEST_TOKEN_SYMBOL, TEST_TOKEN_TOTAL_SUPPLY, TEST_USER2_ID, TEST_USER3_ID, TEST_USER_ENTITY_TYPE, TEST_USER_ID } from "../fixtures";
+import { TEST_ACTIONMETADATA_ENTITY_TYPE, TEST_ACTION_ENTITY_TYPE, TEST_APPFUNGIBLETOKEN_ENTITY_TYPE, TEST_APP_ID, TEST_BADGETOKEN_ENTITY_TYPE, TEST_BADGETOKEN_ID, TEST_MISSIONFUNGIBLETOKEN_ENTITY_TYPE, TEST_MISSIONMETADATA_ENTITY_TYPE, TEST_MISSION_ENTITY_TYPE, TEST_TOKEN_DECIMALS, TEST_TOKEN_ID, TEST_TOKEN_MINTED_ACTION_ID, TEST_TOKEN_MINTED_MISSION_ID, TEST_TOKEN_MINTED_MISSION_URI, TEST_TOKEN_MINTED_URI, TEST_TOKEN_NAME, TEST_TOKEN_SYMBOL, TEST_TOKEN_TOTAL_SUPPLY, TEST_USER2_ID, TEST_USER3_ID, TEST_USER_ENTITY_TYPE, TEST_USER_ID } from "../fixtures";
 import { ActionId, AppFungibleTokenId, BadgeId, MissionId, ZERO_ADDRESS, loadBadgeToken } from "../../src/helpers";
 import { Address, BigInt, ethereum } from "@graphprotocol/graph-ts";
 import { Mission } from "../../generated/schema";
@@ -31,7 +31,7 @@ describe("RewardsFacet tests", () => {
 
         assert.fieldEquals(TEST_ACTIONMETADATA_ENTITY_TYPE, actionId, "id", actionId);
         assert.fieldEquals(TEST_ACTIONMETADATA_ENTITY_TYPE, actionId, "name", TEST_TOKEN_MINTED_ACTION_ID);
-        assert.fieldEquals(TEST_ACTIONMETADATA_ENTITY_TYPE, actionId, "URI", TEST_TOKEN_MINTED_URI);
+        assert.fieldEquals(TEST_ACTIONMETADATA_ENTITY_TYPE, actionId, "URI", TEST_TOKEN_MINTED_MISSION_URI);
     })
 
     test("Token minted MISSION", () => {
@@ -56,7 +56,7 @@ describe("RewardsFacet tests", () => {
 
         assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "id", missionId);
         assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "name", TEST_TOKEN_MINTED_MISSION_ID);
-        assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "URI", TEST_TOKEN_MINTED_URI);
+        assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "URI", TEST_TOKEN_MINTED_MISSION_URI);
 
         assert.fieldEquals(TEST_MISSIONFUNGIBLETOKEN_ENTITY_TYPE, missionFungibleTokenId, "id", missionFungibleTokenId);
         assert.fieldEquals(TEST_MISSIONFUNGIBLETOKEN_ENTITY_TYPE, missionFungibleTokenId, "amount_rewarded", event.params.amount.toString());
@@ -87,7 +87,7 @@ describe("RewardsFacet tests", () => {
 
         assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "id", missionId);
         assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "name", TEST_TOKEN_MINTED_MISSION_ID);
-        assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "URI", TEST_TOKEN_MINTED_URI);
+        assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "URI", TEST_TOKEN_MINTED_MISSION_URI);
 
         assert.fieldEquals(TEST_MISSIONFUNGIBLETOKEN_ENTITY_TYPE, missionFungibleTokenId, "id", missionFungibleTokenId);
         assert.fieldEquals(TEST_MISSIONFUNGIBLETOKEN_ENTITY_TYPE, missionFungibleTokenId, "amount_rewarded", event.params.amount.toString());
@@ -132,7 +132,7 @@ describe("RewardsFacet tests", () => {
 
             assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "id", missionId);
             assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "name", TEST_TOKEN_MINTED_MISSION_ID);
-            assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "URI", TEST_TOKEN_MINTED_URI);
+            assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "URI", TEST_TOKEN_MINTED_MISSION_URI);
 
             assert.fieldEquals(TEST_MISSIONFUNGIBLETOKEN_ENTITY_TYPE, missionFungibleTokenId, "id", missionFungibleTokenId);
             assert.fieldEquals(TEST_MISSIONFUNGIBLETOKEN_ENTITY_TYPE, missionFungibleTokenId, "amount_rewarded", event.params.amount.toString());
@@ -164,7 +164,7 @@ describe("RewardsFacet tests", () => {
 
             assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "id", missionId);
             assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "name", TEST_TOKEN_MINTED_MISSION_ID);
-            assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "URI", TEST_TOKEN_MINTED_URI);
+            assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "URI", TEST_TOKEN_MINTED_MISSION_URI);
 
             assert.fieldEquals(TEST_MISSIONFUNGIBLETOKEN_ENTITY_TYPE, missionFungibleTokenId, "id", missionFungibleTokenId);
             assert.fieldEquals(TEST_MISSIONFUNGIBLETOKEN_ENTITY_TYPE, missionFungibleTokenId, "amount_rewarded", event.params.amount.toString());
@@ -201,7 +201,7 @@ describe("RewardsFacet tests", () => {
 
         assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "id", missionId);
         assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "name", TEST_TOKEN_MINTED_MISSION_ID);
-        assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "URI", TEST_TOKEN_MINTED_URI);
+        assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "URI", TEST_TOKEN_MINTED_MISSION_URI);
 
         const quantity = event.params.quantity;
         for (let i = 0; i < quantity.toI32(); i++) {
@@ -241,7 +241,7 @@ describe("RewardsFacet tests", () => {
 
         assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "id", missionId);
         assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "name", TEST_TOKEN_MINTED_MISSION_ID);
-        assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "URI", TEST_TOKEN_MINTED_URI);
+        assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "URI", TEST_TOKEN_MINTED_MISSION_URI);
 
         const quantity = event.params.quantity;
         for (let i = 0; i < quantity.toI32(); i++) {
@@ -285,7 +285,7 @@ describe("RewardsFacet tests", () => {
 
         assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "id", missionId);
         assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "name", TEST_TOKEN_MINTED_MISSION_ID);
-        assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "URI", TEST_TOKEN_MINTED_URI);
+        assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "URI", TEST_TOKEN_MINTED_MISSION_URI);
 
         const quantity = event.params.quantity;
         for (let i = 0; i < quantity.toI32(); i++) {
@@ -325,7 +325,7 @@ describe("RewardsFacet tests", () => {
 
         assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "id", missionId);
         assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "name", TEST_TOKEN_MINTED_MISSION_ID);
-        assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "URI", TEST_TOKEN_MINTED_URI);
+        assert.fieldEquals(TEST_MISSIONMETADATA_ENTITY_TYPE, missionId, "URI", TEST_TOKEN_MINTED_MISSION_URI);
 
         const mission = Mission.load(missionId);
         assert.assertTrue(mission!.badges != null, "Mission has badges");